optimaize match process

+++ biu/BLForgotPasswordViewController.m
@@ -133,7 +133,7 @@ static NSInteger const BL_RESET_SUCCESS_ALERTVIEW = 1;
     
     [[BLHTTPClient sharedBLHTTPClient] passcode:_code phoneNumber:_tfPhone.text success:^(NSURLSessionDataTask *task, id responseObject) {
         _secondLeftToResend = 60;
-        _lbSecondLeft.text = [NSString stringWithFormat:@"%ld", _secondLeftToResend];
+        _lbSecondLeft.text = [NSString stringWithFormat:@"%ld", (long)_secondLeftToResend];
         _btnGetCode.enabled = NO;
         [self showSecondToResend];
         _timer = [NSTimer scheduledTimerWithTimeInterval:1 target:self selector:@selector(timerfired) userInfo:nil repeats:YES];
